requirements.txt
robin_stocks
tdameritrade~=0.1.1
pandas~=1.1.3
selenium~=3.141.0
python-dateutil~=2.8.1
pytz~=2020.1
bintrees~=2.1.0
numpy~=1.19.2
requests~=2.24.0
tda-api~=0.7.1